Flutter: start implementing Login
Use additional libraries:
1. requests - this library make requests with session cookies(like browser)
So it keeps track of login.
2. shared_preferences - needed for the previous library as it stores cookie.

For now, I implemented login button. When user enters 'admin' 'admin'
it sends login request and then sends request to get loggined user name and
prints it to the console('admin' in this case).
Next step will be to display this in Navbar, maybe permanently...